--- src/main/java/Need/Timer_setup.java
@@ -1,15 +1,0 @@
-import entities.Timertomato;
-
-public class Timer_setup {
-
-    public void startWorkTimer(){
-        String timer = Timertomato.getWorkTime();
-        //timer.start();
-    }
-
-    public void startRestTimer(){
-        String timer = Timertomato.getRestTime();
-        //timer.start();
-    }
-
-}
