+++ src/main/java/UI/ButtonforView.java
@@ -1,4 +1,4 @@
-package UI;
+package ui;
 
 import javax.swing.*;
