+++ src/main/java/UI/ColorforView.java
@@ -1,4 +1,4 @@
-package UI;
+package ui;
 
 import java.awt.*;
