+++ src/main/java/UI/TimerScreen.java
@@ -1,4 +1,4 @@
-package UI;
+package ui;
 
 import net.miginfocom.swing.MigLayout;
