--- src/main/java/example_user/UserExample.java
@@ -1,30 +1,0 @@
-package example_user;
-
-import entities.*;
-
-public class UserExample {
-
-    public static HabitFactory hFactory = new HabitFactory();
-
-    static TaskFactory tFactory = new TaskFactory();
-
-    static CategoryFactory cFactory =  new CategoryFactory();
-
-    static Timer t = new Timer();
-
-    static User person = new User("hello123@gmail.com", "12345678", "Harry", "Potter",
-            hFactory, tFactory, cFactory, t);
-
-    public static User sendUser(){
-        return person;
-    }
-
-    public static void send(){
-        System.out.println(person.getHabitCollection().habitCollection.toString());
-    }
-
-    public static void main(String[] args) {
-        javax.swing.SwingUtilities.invokeLater(UserExample::send);
-    }
-
-}
