--- src/main/java/use_cases/Example.java
@@ -1,4 +1,0 @@
-package use_cases;
-
-public class Example {
-}
