+++ src/main/java/use_cases/timer_needed_use_cases/String_to_int.java
@@ -1,3 +1,5 @@
+package use_cases.timer_needed_use_cases;
+
 import java.util.Scanner;
 public class String_to_int {
     public static void main(String[] args) {
