+++ src/main/java/use_cases/timer_needed_use_cases/TimerView_FixButton.java
@@ -1,5 +1,6 @@
-package Need;
+package use_cases.timer_needed_use_cases;
 
+import entities.TimerTomato;
 import net.miginfocom.swing.MigLayout;
 
 import javax.swing.*;
@@ -9,36 +10,34 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
 @SuppressWarnings("serial")
-public class TimerView extends JFrame
+public class TimerView_FixButton extends JFrame
 {
     private JPanel topPanel;
     String number1;
     String number2;
     // Custom Components
-    // RGB Codes for Blue. Used as Background.
-    private final Color Blue = new Color(0, 50, 200);
+    // RGB Codes for skyBlue. Used as Background.
+    private final Color skyBlue = new Color(0, 50, 200);
 
     // Used for displaying the timer values.
-    private final Font timerStyle = new Font("Good", Font.BOLD, 140);
+    private final Font timerStyle = new Font("MonoAlphabet", Font.BOLD, 140);
 
-    // RGB codes for Red. Used as Background for buttons.
-    private final Color Red = new Color(190, 30, 50);
-
-    // Used in the buttons in the Pomodoro Timer section.
-    private final Font formBTStyles = new Font("formBT", Font.BOLD, 20);
+    // RGB codes for cardinalRed. Used as Background for buttons.
+    private final Color cardinalRed = new Color(189, 32, 49);
 
     // Used in the buttons in the Pomodoro Timer section.
-    private final Font delayLabelStyles = new Font("delaylabel", Font.ITALIC, 20);
+    private final Font formBTStyles = new Font("Existence", Font.BOLD, 16);
 
-    // RGB code for Green. Used as background when the timer is paused.
-    private final Color Green = new Color(20, 140, 30);
+    // Used in the buttons in the Pomodoro Timer section.
+    private final Font delayLabelStyles = new Font("Existence", Font.ITALIC, 20);
+
+    // RGB code for indiaGreen. Used as background when the timer is paused.
+    private final Color indiaGreen = new Color(19, 136, 8);
 
     private static final int ORIGINAL_COUNTDOWN_MINUTES = 25;
     private static final int ORIGINAL_COUNTDOWN_SECONDS = 0;
-
     private int[] workTimeMin = {-1};
     private int[] restTimeMin = {-1};
-
     private static final int ORIGINAL_SHORTBREAK_MINUTES = 5;
     private static final int ORIGINAL_SHORTBREAK_SECONDS = 0;
     private static final int ORIGINAL_LONGBREAK_MINUTES = 15;
@@ -52,41 +51,36 @@ public class TimerView extends JFrame
     private JLabel separator;
     private JLabel secondLabel;
     private JLabel delayRemainingLabel;
-
     private JButton startPauseBT;
     private JButton stopBT;
     private JButton continueBT;
-
     private Icon startIcon;
     private Icon pauseIcon;
     private Icon stopIcon;
     private Icon skipIcon;
-
     private Timer countDown;
-    private Timer workTimer;
-    private Timer restTimer;
+    private Timer shortTimer;
+    private Timer longTimer;
     private Timer delayTimer;
-
     private int secondsRemaining;
     private int minutesRemaining;
     private int delayRemaining;
     private int roundsCompleted; // No. of Pomodoro rounds completed.
 
-    //private static final String workTime = Timertomato.getWorkTime();
-    //static String[] work = workTime.split("\\:");
+    private static final String workTime = TimerTomato.getWorkTime();
+    private static final String restTime = TimerTomato.getRestTime();
+    private static int workTime_minutes;
+    private static int workTime_seconds;
+    private static int restTimer_minutes;
+    private static int restTimer_seconds;
 
-    //private static final int workTime_minutes = Integer.parseInt(work[0]);
-    //private static final int workTime_seconds = Integer.parseInt(work[1]);
-
-    //private static final String restTime = Timertomato.getRestTime();
-    //static String[] rest = getRestTime().split("\\:");
-    //private static final int restTimer_minutes = Integer.parseInt(rest[0]);
-    //private static final int restTimer_seconds = Integer.parseInt(rest[1]);
+    private Timer workTimer;
+    private Timer restTimer;
 
 
-    public TimerView()
+    public void TimerView_FixButton()
     {
-        setTitle("Tomato Timer");
+        setTitle("Pomodoro Timer");
 
         // Setting up the Cloud Look and Feel of the GUI application
         try
@@ -113,7 +107,7 @@ public class TimerView extends JFrame
         }
 
         topPanel = new JPanel();
-        topPanel.setPreferredSize(new Dimension(900, 600));
+        topPanel.setPreferredSize(new Dimension(900, 594));
         topPanel.setLayout(new BorderLayout());
         getContentPane().add(topPanel, BorderLayout.CENTER);
         roundsCompleted = 0;
@@ -125,7 +119,7 @@ public class TimerView extends JFrame
     private JPanel addMainTimer()
     {
         timerPanel = new JPanel();
-        timerPanel.setBackground(Blue);
+        timerPanel.setBackground(skyBlue);
         timerPanel.setLayout(new MigLayout("insets 115 0 0 0", "", "[][]0[]"));
 
         minuteLabel = new JLabel(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
@@ -143,21 +137,22 @@ public class TimerView extends JFrame
         secondLabel.setFont(timerStyle);
         timerPanel.add(secondLabel, "alignx center, height 145!, wrap");
 
-        // startIcon = new ImageIcon("src/Play.gif");
-        // pauseIcon = new ImageIcon("src/Pause.png");
+        //startIcon = new ImageIcon("src/Play.png");
+        //pauseIcon = new ImageIcon("src/Pause.png");
+        //startPauseBT = new JButton("Start", startIcon);
         startPauseBT = new JButton("Start");
-        startPauseBT.setContentAreaFilled(true);
-        startPauseBT.setBackground(Green);
+        startPauseBT.setContentAreaFilled(false);
+        startPauseBT.setBackground(indiaGreen);
         startPauseBT.setActionCommand("Begin");
         startPauseBT.setForeground(Color.white);
         startPauseBT.setFont(formBTStyles);
-        timerPanel.add(startPauseBT, "gaptop 5, alignx center, split 3, spanx, pushx");
+        timerPanel.add(startPauseBT, "gaptop 10, alignx center, split 3, spanx, pushx");
 
         //skipIcon = new ImageIcon("src/Skip.png");
         //continueBT = new JButton("Skip", skipIcon);
         continueBT = new JButton("Skip");
-        continueBT.setBackground(Red);
-        continueBT.setContentAreaFilled(true);
+        continueBT.setBackground(cardinalRed);
+        continueBT.setContentAreaFilled(false);
         continueBT.setForeground(Color.white);
         continueBT.setFont(formBTStyles);
         continueBT.setVisible(false);
@@ -167,12 +162,12 @@ public class TimerView extends JFrame
         //stopIcon = new ImageIcon("src/Stop.png");
         //stopBT = new JButton("Stop", stopIcon);
         stopBT = new JButton("Stop");
-        stopBT.setContentAreaFilled(true);
-        stopBT.setBackground(Red);
+        stopBT.setContentAreaFilled(false);
+        stopBT.setBackground(cardinalRed);
         stopBT.setActionCommand("Stop");
         stopBT.setForeground(Color.white);
         stopBT.setFont(formBTStyles);
-        //timerPanel.add(stopBT, "gaptop 5, alignx center, split 3, spanx, pushx");
+        //timerPanel.add(stopBT, "gaptop 10, alignx center, split 3, spanx, pushx");
         timerPanel.add(stopBT);
 
         delayRemainingLabel = new JLabel("Resumes in " + TOTAL_DELAY_TIME + " seconds");
@@ -180,42 +175,34 @@ public class TimerView extends JFrame
         delayRemainingLabel.setVisible(false);
         delayRemainingLabel.setFont(delayLabelStyles);
         timerPanel.add(delayRemainingLabel, "alignx center");
-
-        JTextField workText = new JTextField(100);
-        workText.setPreferredSize(new Dimension(1,50));
-        timerPanel.add(workText, "alignx center, split 3");
-        JTextField restText = new JTextField(100);
-        restText.setPreferredSize(new Dimension(1,50));
-        timerPanel.add(restText, "alignx center, split 3");
-        JButton SettingButton = new JButton("Setting");
-        SettingButton.setFont(new Font("", Font.PLAIN, 30));
-        timerPanel.add(SettingButton, "alignx center, split 3");
-
-
-        SettingButton.addActionListener(new ActionListener(){
+        JTextField text3 = new JTextField(100);
+        text3.setPreferredSize(new Dimension(1,50));
+        timerPanel.add(text3);
+        JTextField text4 = new JTextField(100);
+        text4.setPreferredSize(new Dimension(1,50));
+        timerPanel.add(text4);
+        JButton button1 = new JButton("Setting");
+        button1.setFont(new Font("", Font.PLAIN, 50));
+        timerPanel.add(button1);
+        button1.addActionListener(new ActionListener(){
 
             @Override
             public void actionPerformed(ActionEvent e) {
                 // set time
-                number1 = workText.getText();
-                number2 = restText.getText();
+                number1 = text3.getText();
+                number2 = text4.getText();
                 int i1 = Integer.parseInt(number1);
-                workTimeMin[0] = i1;
                 int i2 = Integer.parseInt(number2);
-                restTimeMin[0] = i2;
                 minuteLabel.setText(String.format("%02d", i1));
                 secondLabel.setText(String.format("%02d", i2));
-                runMainTimer();
             }
 
         });
-
         startPauseBT.addActionListener((ActionEvent event) -> {
             switch (event.getActionCommand()) {
                 case "Pause":
                     countDown.stop();
                     countDownPaused();
-                    SettingButton.setVisible(false);
                     startPauseBT.setText("Continue");
                     startPauseBT.setActionCommand("Begin");
                     break;
@@ -223,14 +210,12 @@ public class TimerView extends JFrame
                     countDown.start();
                     stopBT.setEnabled(true);
                     //startPauseBT.setIcon(pauseIcon);
-                    SettingButton.setVisible(false);
                     startPauseBT.setText("Pause");
                     startPauseBT.setActionCommand("Pause");
                 case "Begin": // Continue
                     countDown.start();
                     stopBT.setEnabled(true);
                     //startPauseBT.setIcon(pauseIcon);
-                    SettingButton.setVisible(false);
                     startPauseBT.setText("Pause");
                     startPauseBT.setActionCommand("Pause");
                     break;
@@ -240,14 +225,14 @@ public class TimerView extends JFrame
         });
 
         continueBT.addActionListener((ActionEvent event) -> {
-            if(event.getActionCommand().equals("SkipRestTimer"))
+            if(event.getActionCommand().equals("SkipShortTimer"))
             {
-                restTimer.stop();
+                shortTimer.stop();
                 roundsCompleted++;
             }
-            else if(event.getActionCommand().equals("SkipWorkTimer"))
+            else if(event.getActionCommand().equals("SkipLongTimer"))
             {
-                workTimer.stop();
+                longTimer.stop();
                 roundsCompleted = 0; // Timer is Reset.
             }
 
@@ -260,15 +245,12 @@ public class TimerView extends JFrame
         stopBT.addActionListener((ActionEvent event) -> {
             countDown.stop();
             secondsRemaining = ORIGINAL_COUNTDOWN_SECONDS;
-            // minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
-            minutesRemaining = workTimeMin[0];
-            // minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
-            minuteLabel.setText(String.format("%02d", workTimeMin[0]));
+            minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
+            minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
             secondLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_SECONDS));
             stopBT.setEnabled(false);
             roundsCompleted = 0; // Timer is Reset.
             //startPauseBT.setIcon(startIcon);
-            SettingButton.setVisible(true);
             startPauseBT.setText("Start");
             startPauseBT.setActionCommand("Start");
         });
@@ -281,25 +263,23 @@ public class TimerView extends JFrame
     private void runMainTimer()
     {
         //System.out.println("Start Main " + String.format("%d", roundsCompleted));
-        // minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
-        minutesRemaining = workTimeMin[0];
+        minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
         secondsRemaining = ORIGINAL_COUNTDOWN_SECONDS;
 
-        // minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
-        minuteLabel.setText(String.format("%02d", workTimeMin[0]));
+        minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
         secondLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_SECONDS));
 
         if(roundsCompleted == ONE_POMODORO_CYCLE)
         {
             //System.out.println("Stop Long " + roundsCompleted);
-            workTimer.stop();
+            longTimer.stop();
             roundsCompleted = 0; // Timer is Reset.
         }
 
         else if(roundsCompleted > 0 && roundsCompleted % 2 == 0)
         {
             //System.out.println("Stop Short " + roundsCompleted);
-            restTimer.stop();
+            shortTimer.stop();
         }
         // else
         {
@@ -314,17 +294,17 @@ public class TimerView extends JFrame
                     //startPauseBT.setText("Begin");
                     //startPauseBT.setActionCommand("Start");
                     //mainCompleted = false;
-                    //roundsCompleted++;
+                    roundsCompleted++;
 
                     // Selection of which break timer to run.
-                    //if(roundsCompleted == ONE_POMODORO_CYCLE)
-                    //{
-                       // runWorkTimer();
-                    //}
-                    //else if(roundsCompleted > 0 && roundsCompleted % 2 == 0)
-                    //{
-                        runRestTimer();
-                    //}
+                    if(roundsCompleted == ONE_POMODORO_CYCLE)
+                    {
+                        runLongTimer();
+                    }
+                    else if(roundsCompleted > 0 && roundsCompleted % 2 == 0)
+                    {
+                        runShortTimer();
+                    }
                 }
                 else
                 {
@@ -336,8 +316,7 @@ public class TimerView extends JFrame
             }
             else
             {
-                //if(secondsRemaining > ORIGINAL_COUNTDOWN_SECONDS)
-                if(secondsRemaining > 0)
+                if(secondsRemaining > ORIGINAL_COUNTDOWN_SECONDS)
                 {
                     secondsRemaining -= 1;
                     secondLabel.setText(String.format("%02d", secondsRemaining));
@@ -348,23 +327,21 @@ public class TimerView extends JFrame
         countDown.stop();
     }
 
-    private void runRestTimer()
+    private void runShortTimer()
     {
         //System.out.println("Start Short " + String.format("%d", roundsCompleted));
-        //minutesRemaining = ORIGINAL_SHORTBREAK_MINUTES;
-        minutesRemaining = restTimeMin[0];
+        minutesRemaining = ORIGINAL_SHORTBREAK_MINUTES;
         secondsRemaining = ORIGINAL_SHORTBREAK_SECONDS;
-        stopBT.setVisible(true);
+        stopBT.setVisible(false);
         startPauseBT.setVisible(false);
         continueBT.setVisible(false);
         continueBT.setActionCommand("SkipShortTimer");
         countDown.stop();
 
-        //minuteLabel.setText(String.format("%02d", ORIGINAL_SHORTBREAK_MINUTES));
-        minuteLabel.setText(String.format("%02d", restTimeMin[0]));
+        minuteLabel.setText(String.format("%02d", ORIGINAL_SHORTBREAK_MINUTES));
         secondLabel.setText(String.format("%02d", ORIGINAL_SHORTBREAK_SECONDS));
 
-        restTimer = new Timer(INTERVAL, (ActionEvent event) -> {
+        shortTimer = new Timer(INTERVAL, (ActionEvent event) -> {
             if(secondsRemaining == ORIGINAL_SHORTBREAK_SECONDS)
             {
                 if(minutesRemaining == 0)
@@ -390,24 +367,24 @@ public class TimerView extends JFrame
             }
         });
 
-        restTimer.start();
+        shortTimer.start();
     }
 
-    private void runWorkTimer()
+    private void runLongTimer()
     {
         //System.out.println("Start Long " + String.format("%d", roundsCompleted));
         minutesRemaining = ORIGINAL_LONGBREAK_MINUTES;
         secondsRemaining = ORIGINAL_LONGBREAK_SECONDS;
-        stopBT.setVisible(true);
-        startPauseBT.setVisible(true);
+        stopBT.setVisible(false);
+        startPauseBT.setVisible(false);
         continueBT.setVisible(true);
-        continueBT.setActionCommand("SkipWorkTimer");
+        continueBT.setActionCommand("SkipLongTimer");
         countDown.stop();
 
         minuteLabel.setText(String.format("%02d", ORIGINAL_LONGBREAK_MINUTES));
         secondLabel.setText(String.format("%02d", ORIGINAL_LONGBREAK_SECONDS));
 
-        workTimer = new Timer(INTERVAL, (ActionEvent event) -> {
+        longTimer = new Timer(INTERVAL, (ActionEvent event) -> {
             if(secondsRemaining == 0)
             {
                 if(minutesRemaining == 0)
@@ -433,7 +410,7 @@ public class TimerView extends JFrame
             }
         });
 
-        workTimer.start();
+        longTimer.start();
     }
 
     private void countDownPaused()
@@ -479,7 +456,7 @@ public class TimerView extends JFrame
 
     public static void main(String[] args)
     {
-        SwingUtilities.invokeLater(TimerView::runGUI);
+        SwingUtilities.invokeLater(TimerView_FixButton::runGUI);
     }
 
 
@@ -488,3 +465,4 @@ public class TimerView extends JFrame
 
 
 }
+
