+++ src/main/java/use_cases/timer_needed_use_cases/TimerView_FixCycle.java
@@ -1,6 +1,6 @@
-package Need;
+package use_cases.timer_needed_use_cases;
 
-import entities.Timertomato;
+import entities.TimerTomato;
 import net.miginfocom.swing.MigLayout;
 
 import javax.swing.*;
@@ -10,7 +10,7 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
 @SuppressWarnings("serial")
-public class TimerView_FixButton extends JFrame
+public class TimerView_FixCycle extends JFrame
 {
     private JPanel topPanel;
     String number1;
@@ -34,11 +34,13 @@ public class TimerView_FixButton extends JFrame
     // RGB code for indiaGreen. Used as background when the timer is paused.
     private final Color indiaGreen = new Color(19, 136, 8);
 
-    private static final int ORIGINAL_COUNTDOWN_MINUTES = 25;
+    //private static final int ORIGINAL_COUNTDOWN_MINUTES = 25;
+    private int ORIGINAL_COUNTDOWN_MINUTES = -1;
     private static final int ORIGINAL_COUNTDOWN_SECONDS = 0;
     private int[] workTimeMin = {-1};
     private int[] restTimeMin = {-1};
-    private static final int ORIGINAL_SHORTBREAK_MINUTES = 5;
+    //private static final int ORIGINAL_SHORTBREAK_MINUTES = 5;
+    private int ORIGINAL_SHORTBREAK_MINUTES = -1;
     private static final int ORIGINAL_SHORTBREAK_SECONDS = 0;
     private static final int ORIGINAL_LONGBREAK_MINUTES = 15;
     private static final int ORIGINAL_LONGBREAK_SECONDS = 0;
@@ -67,8 +69,8 @@ public class TimerView_FixButton extends JFrame
     private int delayRemaining;
     private int roundsCompleted; // No. of Pomodoro rounds completed.
 
-    private static final String workTime = Timertomato.getWorkTime();
-    private static final String restTime = Timertomato.getRestTime();
+    private static final String workTime = TimerTomato.getWorkTime();
+    private static final String restTime = TimerTomato.getRestTime();
     private static int workTime_minutes;
     private static int workTime_seconds;
     private static int restTimer_minutes;
@@ -181,7 +183,7 @@ public class TimerView_FixButton extends JFrame
         JTextField text4 = new JTextField(100);
         text4.setPreferredSize(new Dimension(1,50));
         timerPanel.add(text4);
-        JButton button1 = new JButton("Setting");
+        JButton button1 = new JButton("Set Time");
         button1.setFont(new Font("", Font.PLAIN, 50));
         timerPanel.add(button1);
         button1.addActionListener(new ActionListener(){
@@ -192,9 +194,12 @@ public class TimerView_FixButton extends JFrame
                 number1 = text3.getText();
                 number2 = text4.getText();
                 int i1 = Integer.parseInt(number1);
+                workTimeMin[0] = i1;
                 int i2 = Integer.parseInt(number2);
+                restTimeMin[0] = i2;
                 minuteLabel.setText(String.format("%02d", i1));
                 secondLabel.setText(String.format("%02d", i2));
+                runMainTimer();
             }
 
         });
@@ -203,6 +208,7 @@ public class TimerView_FixButton extends JFrame
                 case "Pause":
                     countDown.stop();
                     countDownPaused();
+                    button1.setVisible(false);
                     startPauseBT.setText("Continue");
                     startPauseBT.setActionCommand("Begin");
                     break;
@@ -210,12 +216,14 @@ public class TimerView_FixButton extends JFrame
                     countDown.start();
                     stopBT.setEnabled(true);
                     //startPauseBT.setIcon(pauseIcon);
+                    button1.setVisible(false);
                     startPauseBT.setText("Pause");
                     startPauseBT.setActionCommand("Pause");
                 case "Begin": // Continue
                     countDown.start();
                     stopBT.setEnabled(true);
                     //startPauseBT.setIcon(pauseIcon);
+                    button1.setVisible(false);
                     startPauseBT.setText("Pause");
                     startPauseBT.setActionCommand("Pause");
                     break;
@@ -244,18 +252,24 @@ public class TimerView_FixButton extends JFrame
 
         stopBT.addActionListener((ActionEvent event) -> {
             countDown.stop();
+            shortTimer.stop();
             secondsRemaining = ORIGINAL_COUNTDOWN_SECONDS;
-            minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
-            minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
+            //minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
+            minutesRemaining = workTimeMin[0];
+            //minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
+            minuteLabel.setText(String.format("%02d", workTimeMin[0]));
             secondLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_SECONDS));
             stopBT.setEnabled(false);
             roundsCompleted = 0; // Timer is Reset.
             //startPauseBT.setIcon(startIcon);
+            button1.setVisible(true);
             startPauseBT.setText("Start");
             startPauseBT.setActionCommand("Start");
         });
 
+
         runMainTimer();
+
 
         return timerPanel;
     }
@@ -263,10 +277,12 @@ public class TimerView_FixButton extends JFrame
     private void runMainTimer()
     {
         //System.out.println("Start Main " + String.format("%d", roundsCompleted));
-        minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
+        //minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
+        minutesRemaining = workTimeMin[0];
         secondsRemaining = ORIGINAL_COUNTDOWN_SECONDS;
 
-        minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
+        //minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
+        minuteLabel.setText(String.format("%02d", workTimeMin[0]));
         secondLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_SECONDS));
 
         if(roundsCompleted == ONE_POMODORO_CYCLE)
@@ -287,24 +303,11 @@ public class TimerView_FixButton extends JFrame
         }
 
         countDown = new Timer(INTERVAL, (ActionEvent event) -> {
-            if(secondsRemaining == ORIGINAL_COUNTDOWN_SECONDS)
+            if(secondsRemaining == 0)
             {
                 if(minutesRemaining == 0)
                 {
-                    //startPauseBT.setText("Begin");
-                    //startPauseBT.setActionCommand("Start");
-                    //mainCompleted = false;
-                    roundsCompleted++;
-
-                    // Selection of which break timer to run.
-                    if(roundsCompleted == ONE_POMODORO_CYCLE)
-                    {
-                        runLongTimer();
-                    }
-                    else if(roundsCompleted > 0 && roundsCompleted % 2 == 0)
-                    {
-                        runShortTimer();
-                    }
+                    runShortTimer();
                 }
                 else
                 {
@@ -316,7 +319,7 @@ public class TimerView_FixButton extends JFrame
             }
             else
             {
-                if(secondsRemaining > ORIGINAL_COUNTDOWN_SECONDS)
+                if(secondsRemaining > 0)
                 {
                     secondsRemaining -= 1;
                     secondLabel.setText(String.format("%02d", secondsRemaining));
@@ -330,15 +333,17 @@ public class TimerView_FixButton extends JFrame
     private void runShortTimer()
     {
         //System.out.println("Start Short " + String.format("%d", roundsCompleted));
-        minutesRemaining = ORIGINAL_SHORTBREAK_MINUTES;
+        //minutesRemaining = ORIGINAL_SHORTBREAK_MINUTES;
+        minutesRemaining = restTimeMin[0];
         secondsRemaining = ORIGINAL_SHORTBREAK_SECONDS;
-        stopBT.setVisible(false);
+        stopBT.setVisible(true);
         startPauseBT.setVisible(false);
         continueBT.setVisible(false);
         continueBT.setActionCommand("SkipShortTimer");
         countDown.stop();
 
-        minuteLabel.setText(String.format("%02d", ORIGINAL_SHORTBREAK_MINUTES));
+        //minuteLabel.setText(String.format("%02d", ORIGINAL_SHORTBREAK_MINUTES));
+        minuteLabel.setText(String.format("%02d", restTimeMin[0]));
         secondLabel.setText(String.format("%02d", ORIGINAL_SHORTBREAK_SECONDS));
 
         shortTimer = new Timer(INTERVAL, (ActionEvent event) -> {
@@ -456,7 +461,7 @@ public class TimerView_FixButton extends JFrame
 
     public static void main(String[] args)
     {
-        SwingUtilities.invokeLater(TimerView_FixButton::runGUI);
+        SwingUtilities.invokeLater(TimerView_FixCycle::runGUI);
     }
 
 
@@ -465,4 +470,3 @@ public class TimerView_FixButton extends JFrame
 
 
 }
-
