+++ src/main/java/use_cases/timer_needed_use_cases/Timer_1.java
@@ -1,3 +1,5 @@
+package use_cases.timer_needed_use_cases;
+
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.Scanner;
