+++ src/main/java/use_cases/timer_needed_use_cases/Timer_countdown.java
@@ -1,10 +1,12 @@
+package use_cases.timer_needed_use_cases;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.text.DecimalFormat;
 
-public class Normal_Timer {
+public class Timer_countdown {
 
     JFrame window;
     JLabel counterLabel;
@@ -20,10 +22,10 @@ public class Normal_Timer {
 
     public static void main(String[] args){
 
-        new Normal_Timer();
+        new Timer_countdown();
     }
 
-    public Normal_Timer() {
+    public Timer_countdown() {
 
         window = new JFrame();
         window.setSize(800, 600);
@@ -37,39 +39,62 @@ public class Normal_Timer {
         window.add(counterLabel);
         window.setVisible(true);
 
-        counterLabel.setText("00:00");
+//        second = 0;
+//        simpleTimer();
+//        timer.start();
+
+        // Normal Timer
+        counterLabel.setText("03:00");
         second = 0;
-        minute = 0;
-        normalTimer();
+        minute = 3;
+        countdownTimer();
         timer.start();
     }
 
-    public void normalTimer() {
+//    public void simpleTimer(){
+//
+//        timer = new Timer(1000, new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+                // What happen set the timer
+
+//                second++;
+
+//                counterLabel.setText(""+ second);
+//            }
+//        });
+
+//    }
+
+
+    public void countdownTimer() {
 
         timer = new Timer(1000, new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-
                 // What happen set the timer
 
-                second++;
+                second--;
                 ddSecond = dFormat.format(second);
                 ddMinute = dFormat.format(minute);
 
                 counterLabel.setText(ddMinute+ ":" + ddSecond);
 
-                if(second==60){
-                    second = 0;
-                    minute++;
+                if(second == -1){
+                    second = 59;
+                    minute = - minute++;
 
                     ddSecond = dFormat.format(second);
                     ddMinute = dFormat.format(minute);
                     counterLabel.setText(ddMinute + ":" + ddSecond);
+                }
+
+                if(minute == 0 && second == 0){
+                    timer.stop();
                 }
             }
         });
 
 
     }
-
-    }
+}
