+++ src/main/java/use_cases/timer_needed_use_cases/Timer_setup.java
@@ -1,0 +1,17 @@
+package use_cases.timer_needed_use_cases;
+
+import entities.TimerTomato;
+
+public class Timer_setup {
+
+    public void startWorkTimer(){
+        String timer = TimerTomato.getWorkTime();
+        //timer.start();
+    }
+
+    public void startRestTimer(){
+        String timer = TimerTomato.getRestTime();
+        //timer.start();
+    }
+
+}
