+++ src/main/java/use_cases/timer_needed_use_cases/hour_and_second.java
@@ -1,4 +1,6 @@
-import entities.Timertomato;
+package use_cases.timer_needed_use_cases;
+
+import entities.TimerTomato;
 
 import javax.swing.*;
 import java.awt.*;
@@ -24,7 +26,7 @@ public class hour_and_second {
 
     public static void main(String[] args){
 
-        new hour_and_second(Timertomato.getWorkTime());
+        new hour_and_second(TimerTomato.getWorkTime());
     }
 
     public hour_and_second(String timer) {
