+++ src/main/java/use_cases/timer_use_cases/Example.java
@@ -1,0 +1,4 @@
+package use_cases.timer_use_cases;
+
+public class Example {
+}
