+++ src/main/java/use_cases/timer_use_cases/editTimerInputBoundary.java
@@ -1,4 +1,4 @@
-package use_cases;
+package use_cases.timer_use_cases;
 
 public class editTimerInputBoundary {
 }
